revert files renaming, add userRole const instead of absolute zero
