+++ routes/user.js
@@ -1,12 +1,8 @@
 const express = require("express");
 const router = express.Router();
 
-const {
-  requireSignin,
-  isAuth,
-  isAdmin,
-} = require("../controllers/auth-controller");
-const { findUserById } = require("../controllers/user-controller");
+const { requireSignin, isAuth, isAdmin } = require("../controllers/auth");
+const { findUserById } = require("../controllers/user");
 
 router.get("/profile/:userId", requireSignin, isAuth, isAdmin, (req, res) => {
   res.json({
